+++ src/main/java/de/bund/digitalservice/ris/domain/DocUnitRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
 @Component
-public interface DocUnitRepository extends ReactiveCrudRepository<DocUnit, Integer> {
+public interface DocUnitRepository extends ReactiveCrudRepository<DocUnit, Long> {
 
   @Query("select id, s3path, filetype from DOC_UNIT where filetype = $1")
   Flux<DocUnit> findByFileType(String filetype);
